Fix: Renaming wrong migration helper title import
