+++ specifyweb/specify/migration_utils/migration_helpers/__init__.py
@@ -1,7 +1,7 @@
 from .helper_0002_schema_config_update import MIGRATION_0002_TABLES
 from .helper_0004_stratigraphy_age import MIGRATION_0004_FIELDS, MIGRATION_0004_TABLES
 from .helper_0007_schema_config_update import MIGRATION_0007_FIELDS
-from .helper_0008_schema_config_update import MIGRATION_0008_FIELDS
+from .helper_0008_ageCitations_fix import MIGRATION_0008_FIELDS
 from .helper_0012_add_cojo_to_schema_config import MIGRATION_0012_FIELDS
 from .helper_0013_collectionobjectgroup_parentcog import MIGRATION_0013_FIELDS
 from .helper_0020_add_tectonicunit_to_pc_in_schema_config import MIGRATION_0020_FIELDS
